Bugfix ClinCnvList: totalCNVsize is now of datatype long long, int was to small for certain samples.

+++ src/cppNGS/ClinCnvList.h
@@ -146,7 +146,7 @@ public:
 	void copyMetaData(const ClinCnvList& rhs);
 
 	///Total size of all CNVs in list
-	int totalCNVsize();
+	long long totalCNVsize();
 
 
 	ClinCnvAnalysisType type() const;
